🔒 refactor(auth.py): move SECRET_KEY logic to AuthSettings class to improve code organization and reusability 🔒 refactor(base.py): remove SECRET_KEY field from Settings class since it is now handled by AuthSettings class 🔒 refactor(manager.py): pass CONFIG_DIR to AuthSettings constructor when creating an instance
